+++ docsrc/conf.py
@@ -30,7 +30,7 @@ release = '1.0.0'
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ['sphinx.ext.autodoc']
+extensions = ['sphinx.ext.autodoc', "sphinx_issues"]
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
@@ -59,3 +59,5 @@ html_logo = "_static/icon.png"
 html_theme_options = {
     "display_version": True
 }
+
+issues_github_path = "TheRenegadeCoder/SnakeMD"