+++ docsrc/version-history.rst
@@ -10,13 +10,21 @@ In an effort to keep history of all the documentation
 for SnakeMD, we've included all old versions below
 as follows:
 
-* `v0.3.0 <https://snakemd.therenegadecoder.com/v0.3.0/>`_
+* `v0.4.0 <https://snakemd.therenegadecoder.com/v0.4.0/>`_ 
+    * Added support for Python logging library (:issue:`22`)
+    * Expanded support for strings in the Header, Paragraph, and MDList classes
+    * Fixed an issue where Paragraphs would sometimes render unexpected spaces (:issue:`23`)
+    * Added GitHub links to version history page
+    * Added support for column alignment on tables (:issue:`4`)
+    * Fixed issue where tables sometimes wouldn't pretty print properly (:issue:`5`)
+
+* `v0.3.0 <https://snakemd.therenegadecoder.com/v0.3.0/>`_ [:pr:`21`]
     * Gave documentation a major overhaul
     * Added support for paragraphs in MDList
     * Added is_text() method to Paragraph
     * Fixed issue where punctuation sometimes rendered with an extra space in front
 
-* `v0.2.0 <https://snakemd.therenegadecoder.com/v0.2.0/>`_
+* `v0.2.0 <https://snakemd.therenegadecoder.com/v0.2.0/>`_ [:pr:`17`]
     * Added support for horizontal rules
     * Added automated testing through PyTest and GitHub Actions
     * Added document verification services